:recycle: refactor: name as component
